Move `./src/s3` to `./src/middlewares/s3`
